Increase memory of readonly to 8GB
This is because the readonly instance of prometheus has seen some
out of memory crashes. The instance is at risk of sudden large
amounts of memory being needed if someone runs a big query on
Grafana (maybe they ask for a stat going back for the last 4 weeks).

Cost of the increase is about £35 per month which seems tolerable.

The choice is 4GB, 8GB, 16GB. There is no intermediate sizes so that
is why I went for 8GB.

I also looked at whether the memory for the other prometheus (not
the readonly one) was appropriate. It is currently sitting at 2GB
out of an available 4GB so that is probably appropriate. It also
has no recent out of memory crashes.